[Certif] BSR reorganize tests files
