+++ certif/tests/acceptance/session-details-certification-candidates-test.js
@@ -1,91 +1,159 @@
 import { module, test } from 'qunit';
-import { click, currentURL, visit } from '@ember/test-helpers';
+import { click, currentURL, visit, fillIn } from '@ember/test-helpers';
 import { setupApplicationTest } from 'ember-qunit';
 import moment from 'moment';
 import {
-  createUserWithMembershipAndTermsOfServiceAccepted,
   createScoUserWithMembershipAndTermsOfServiceAccepted,
+  createUserAndMembershipAndTermsOfServiceAccepted,
   authenticateSession,
 } from '../helpers/test-init';
+import { upload } from 'ember-file-upload/test-support';
 
 import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
 
-module('Acceptance | Session Details', function(hooks) {
+module('Acceptance | Session Details Certification Candidates', function(hooks) {
   
   setupApplicationTest(hooks);
   setupMirage(hooks);
+
+  const validLastName = 'MonNom';
+  const validFirstName = 'MonPrenom';
+  const validBirthCity = 'MaVille';
+  const validBirthProvinceCode = '974';
+  const validBirthCountry = 'MonPays';
+  const validBirthdate = '01021990';
   
   hooks.afterEach(function() {
     const notificationMessagesService = this.owner.lookup('service:notifications');
     notificationMessagesService.clearAll();
   });
 
+  module('When user is not logged in', function() {
+
+    test('it should not be accessible by an unauthenticated user', async function(assert) {
+      const session = server.create('session');
+      
+      // when
+      await visit(`/sessions/${session.id}/candidats`);
+
+      // then
+      assert.equal(currentURL(), '/connexion');
+    });
+  });
+
   module('when user is logged in', function(hooks) {
 
     let user;
+    let certificationCenter;
+    let session;
 
     hooks.beforeEach(async () => {
-      user = createUserWithMembershipAndTermsOfServiceAccepted();
+      ({ user, certificationCenter } = createUserAndMembershipAndTermsOfServiceAccepted());
+      session = server.create('session', { certificationCenterId: certificationCenter.id });
       await authenticateSession(user.id);
     });
-    
-    module('when visiting the candidates tab', function() {
 
-      module('when there is no candidats yet', function() {
+    module('when there is no candidats yet', function() {
 
-        let sessionWithoutCandidates;
+      test('it should display a download button and upload button', async function(assert) {
+        // when
+        await visit(`/sessions/${session.id}/candidats`);
 
-        hooks.beforeEach(function() {
-          sessionWithoutCandidates = server.create('session');
-        });
-
-        test('it should display a download button and upload button', async function(assert) {
-          // when
-          await visit(`/sessions/${sessionWithoutCandidates.id}/candidats`);
-
-          // then
-          assert.dom('[data-test-id="attendance_sheet_download_button"]').exists();
-          assert.dom('[data-test-id="attendance_sheet_download_button"]').hasText('Télécharger (.ods)');
-          assert.dom('[data-test-id="attendance_sheet_upload_button"]').exists();
-          assert.dom('[data-test-id="attendance_sheet_upload_button"]').hasText('Importer (.ods)');
-        });
-
-        test('it should display a sentence when there is no certification candidates yet', async function(assert) {
-          // when
-          await visit(`/sessions/${sessionWithoutCandidates.id}/candidats`);
-
-          // then
-          assert.dom('table tbody').doesNotExist();
-          assert.dom('.table__empty').hasText('En attente de candidats');
-        });
+        // then
+        assert.dom('[data-test-id="attendance_sheet_download_button"]').exists();
+        assert.dom('[data-test-id="attendance_sheet_download_button"]').hasText('Télécharger (.ods)');
+        assert.dom('[data-test-id="attendance_sheet_upload_button"]').exists();
+        assert.dom('[data-test-id="attendance_sheet_upload_button"]').hasText('Importer (.ods)');
       });
 
-      module('when there are few candidats', function() {
+      test('it should display a sentence when there is no certification candidates yet', async function(assert) {
+        // when
+        await visit(`/sessions/${session.id}/candidats`);
 
-        let sessionWithCandidates;
-        let candidates;
+        // then
+        assert.dom('table tbody').doesNotExist();
+        assert.dom('.table__empty').hasText('En attente de candidats');
+      });
+    });
 
-        hooks.beforeEach(function() {
-          sessionWithCandidates = server.create('session');
-          candidates = server.createList('certification-candidate', 3, { sessionId: sessionWithCandidates.id });
+    module('when there are few candidats', function() {
+
+      let sessionWithCandidates;
+      let candidates;
+
+      hooks.beforeEach(function() {
+        sessionWithCandidates = server.create('session', { certificationCenterId: certificationCenter.id });
+        candidates = server.createList('certification-candidate', 3, { sessionId: sessionWithCandidates.id, isLinked: false });
+      });
+
+      test('it should display the list of certification candidates', async function(assert) {
+        // given
+        const aCandidate = candidates[0];
+
+        // when
+        await visit(`/sessions/${sessionWithCandidates.id}/candidats`);
+    
+        // then
+        assert.dom('table tbody tr').exists({ count: 3 });
+        assert.dom(`[data-test-id="panel-candidate__lastName__${aCandidate.id}"]`).hasText(`${aCandidate.lastName}`);
+        assert.dom(`[data-test-id="panel-candidate__firstName__${aCandidate.id}"]`).hasText(`${aCandidate.firstName}`);
+        assert.dom(`[data-test-id="panel-candidate__birthdate__${aCandidate.id}"]`).hasText(`${moment(aCandidate.birthdate, 'YYYY-MM-DD').format('DD/MM/YYYY')}`);
+        assert.dom(`[data-test-id="panel-candidate__birthCity__${aCandidate.id}"]`).hasText(`${aCandidate.birthCity}`);
+        assert.dom(`[data-test-id="panel-candidate__birthProvinceCode__${aCandidate.id}"]`).hasText(`${aCandidate.birthProvinceCode}`);
+        assert.dom(`[data-test-id="panel-candidate__birthCountry__${aCandidate.id}"]`).hasText(`${aCandidate.birthCountry}`);
+        assert.dom(`[data-test-id="panel-candidate__email__${aCandidate.id}"]`).hasText(`${aCandidate.email}`);
+        assert.dom(`[data-test-id="panel-candidate__externalId__${aCandidate.id}"]`).hasText(`${aCandidate.externalId}`);
+      });
+
+      module('on import', function() {
+
+        test('it should replace the candidates list with the imported ones', async function(assert) {
+          // given
+          await visit(`/sessions/${sessionWithCandidates.id}/candidats`);
+          const file = new File(['foo'], `${sessionWithCandidates.id}.addTwoCandidates`);
+  
+          // when
+          await upload('#upload-attendance-sheet', file);
+  
+          // then
+          assert.dom('table tbody tr').exists({ count: 2 });
         });
 
-        test('it should display the list of certification candidates', async function(assert) {
+        test('it should display a success message when uploading a valid file', async function(assert) {
           // given
-          const aCandidate = candidates[0];
+          await visit(`/sessions/${sessionWithCandidates.id}/candidats`);
+          const file = new File(['foo'], 'valid-file');
+          
+          // when
+          await upload('#upload-attendance-sheet', file);
+  
+          // then
+          assert.dom('[data-test-notification-message="success"]').hasText('La liste des candidats a été importée avec succès.');
+        });
+
+        test('it should display the error message when uploading an invalid file', async function(assert) {
+          // given
+          await visit(`/sessions/${sessionWithCandidates.id}/candidats`);
+          const file = new File(['foo'], 'invalid-file');
 
           // when
-          await visit(`/sessions/${sessionWithCandidates.id}/candidats`);
-      
+          await upload('#upload-attendance-sheet', file);
+
           // then
-          assert.dom(`[data-test-id="panel-candidate__lastName__${aCandidate.id}"]`).hasText(`${aCandidate.lastName}`);
-          assert.dom(`[data-test-id="panel-candidate__firstName__${aCandidate.id}"]`).hasText(`${aCandidate.firstName}`);
-          assert.dom(`[data-test-id="panel-candidate__birthdate__${aCandidate.id}"]`).hasText(`${moment(aCandidate.birthdate, 'YYYY-MM-DD').format('DD/MM/YYYY')}`);
-          assert.dom(`[data-test-id="panel-candidate__birthCity__${aCandidate.id}"]`).hasText(`${aCandidate.birthCity}`);
-          assert.dom(`[data-test-id="panel-candidate__birthProvinceCode__${aCandidate.id}"]`).hasText(`${aCandidate.birthProvinceCode}`);
-          assert.dom(`[data-test-id="panel-candidate__birthCountry__${aCandidate.id}"]`).hasText(`${aCandidate.birthCountry}`);
-          assert.dom(`[data-test-id="panel-candidate__email__${aCandidate.id}"]`).hasText(`${aCandidate.email}`);
-          assert.dom(`[data-test-id="panel-candidate__externalId__${aCandidate.id}"]`).hasText(`${aCandidate.externalId}`);
+          assert.dom('[data-test-notification-message="error"]').hasText('Aucun candidat n’a été importé. Une erreur personnalisée Veuillez modifier votre fichier et l’importer à nouveau.');
+        });
+
+        test('it should display a specific error message when importing is forbidden', async function(assert) {
+          // given
+          await visit(`/sessions/${sessionWithCandidates.id}/candidats`);
+          const file = new File(['foo'], 'forbidden-import');
+
+          // when
+          await upload('#upload-attendance-sheet', file);
+
+          // then
+          assert.dom('[data-test-notification-message="error"]')
+            .hasText('La session a débuté, il n\'est plus possible de modifier la liste des candidats.');
         });
 
         test('it should display a warning when the import is not allowed', async function(assert) {
@@ -99,18 +167,79 @@ module('Acceptance | Session Details', function(hooks) {
           assert.dom('.panel-actions__warning').hasText(
             'La session a débuté, vous ne pouvez plus importer une liste de candidats.Si vous souhaitez modifier la liste, vous pouvez ajouter un candidat directement dans le tableau ci-dessous.');
         });
-        
-        module('when adding a candidate in staging for saving', () => {
-  
-          test('it should add a new line when we click on "Ajouter un candidat" button', async function(assert) {
-            // when
+      });
+
+      module('add candidate', () => {
+
+        module('when user is not SCO', function(hooks) {
+
+          hooks.beforeEach(async function() {
             await visit(`/sessions/${sessionWithCandidates.id}/candidats`);
-            await click('[data-test-id="add-certification-candidate-staging__button"]');
-  
-            // then
-            assert.dom('[data-test-id="panel-candidate__lastName__add-staging"]').exists();
           });
-  
+
+          module('when candidate data not valid', function() {
+
+            test('it should leave the line up for modification', async function(assert) {
+              this.server.post('/sessions/:id/certification-candidates', () => ({
+                errors: ['Invalid data'],
+              }), 400);
+              // when
+              await click('[data-test-id="add-certification-candidate-staging__button"]');
+              await _fillFormWithCorrectData();
+              await click('[data-test-id="panel-candidate__action__save"]');
+
+              // then
+              assert.dom('[data-test-id="panel-candidate__lastName__add-staging"]').exists();
+            });
+
+            test('it should display notification error', async function(assert) {
+              this.server.post('/sessions/:id/certification-candidates', () => ({
+                errors: ['Invalid data'],
+              }), 400);
+              // when
+              await click('[data-test-id="add-certification-candidate-staging__button"]');
+              await _fillFormWithCorrectData();
+              await click('[data-test-id="panel-candidate__action__save"]');
+
+              // then
+              assert.dom('[data-test-notification-message="error"]').hasText('Une erreur s\'est produite lors de l\'ajout du candidat.');
+            });
+          });
+
+          module('when candidate data is valid', function() {
+
+            test('it remove the editable line', async function(assert) {
+              // when
+              await click('[data-test-id="add-certification-candidate-staging__button"]');
+              await _fillFormWithCorrectData();
+              await click('[data-test-id="panel-candidate__action__save"]');
+
+              // then
+              assert.dom('[data-test-id="panel-candidate__lastName__add-staging"]').doesNotExist();
+            });
+
+            test('it should display notification success', async function(assert) {
+              // when
+              await click('[data-test-id="add-certification-candidate-staging__button"]');
+              await _fillFormWithCorrectData();
+              await click('[data-test-id="panel-candidate__action__save"]');
+
+              // then
+              assert.dom('[data-test-notification-message="success"]').hasText('Le candidat a été ajouté avec succès.');
+            });
+
+            test('it should add a new candidate entry', async function(assert) {
+              // given
+              // when
+              await click('[data-test-id="add-certification-candidate-staging__button"]');
+              await _fillFormWithCorrectData();
+              await click('[data-test-id="panel-candidate__action__save"]');
+
+              // then
+              assert.dom('table tbody tr').exists({ count: 4 });
+            });
+          });
+
           test('it should remove the line when clicking on cancel button', async function(assert) {
             // when
             await visit(`/sessions/${sessionWithCandidates.id}/candidats`);
@@ -122,44 +251,77 @@ module('Acceptance | Session Details', function(hooks) {
             assert.dom('[data-test-id="panel-candidate__lastName__add-staging"]').doesNotExist();
           });
         });
-      });
 
-      module('prescription sco toogle', function() {
-        let session;
-        let scoUser;
-        const linkToCandidate = '.session-details-controls__navbar-tabs a:nth-of-type(2)';
+        module('when user is SCO', function() {
 
-        hooks.beforeEach(async () => {
-          session = server.create('session');
-          scoUser = createScoUserWithMembershipAndTermsOfServiceAccepted();
+          test('it should display the list of students for session', async function(assert) {
+            // given
+            server.create('feature-toggle', { id: 0, certifPrescriptionSco: true });
+            certificationCenter.update({ type: 'SCO' });
+            server.createList('student', 10);
+
+            // when
+            await visit(`/sessions/${session.id}/candidats`);
+            await click('.button.button--link');
+
+            // then
+            assert.equal(currentURL(), `/sessions/${session.id}/ajout-eleves`);
+            assert.dom('table tbody tr').exists({ count: 10 });
+          });
         });
 
-        [
-          { isFeatureToggleEnabled: false, isUserSco: false },
-          { isFeatureToggleEnabled: true, isUserSco: false },
-          { isFeatureToggleEnabled: false, isUserSco: true },
-          { isFeatureToggleEnabled: true, isUserSco: true },
-        ].forEach(({ isFeatureToggleEnabled, isUserSco }) => {
-          module(`when certification prescription sco feature toggle is ${isFeatureToggleEnabled ? 'enabled' : 'disabled'} and the user is ${isUserSco ? 'SCO' : 'not SCO'}`, () => {
-      
-            test('it should redirect to the default candidates detail view', async (assert) => {
-              // given
-              server.create('feature-toggle', {
-                certifPrescriptionSco: isFeatureToggleEnabled,
-              });
-              const connectedUserId = isUserSco ? scoUser.id : user.id;
-              await authenticateSession(connectedUserId);
-      
-              // when
-              await visit(`/sessions/${session.id}`);
-              await click(linkToCandidate);
-      
-              // then
-              assert.equal(currentURL(), `/sessions/${session.id}/candidats`);
+      });
+    });
+
+    module('prescription sco toogle', function() {
+      let session;
+      let scoUser;
+      const linkToCandidate = '.session-details-controls__navbar-tabs a:nth-of-type(2)';
+
+      hooks.beforeEach(async () => {
+        session = server.create('session');
+        scoUser = createScoUserWithMembershipAndTermsOfServiceAccepted();
+      });
+
+      [
+        { isFeatureToggleEnabled: false, isUserSco: false },
+        { isFeatureToggleEnabled: true, isUserSco: false },
+        { isFeatureToggleEnabled: false, isUserSco: true },
+        { isFeatureToggleEnabled: true, isUserSco: true },
+      ].forEach(({ isFeatureToggleEnabled, isUserSco }) => {
+        module(`when certification prescription sco feature toggle is ${isFeatureToggleEnabled ? 'enabled' : 'disabled'} and the user is ${isUserSco ? 'SCO' : 'not SCO'}`, () => {
+
+          test('it should redirect to the default candidates detail view', async (assert) => {
+            // given
+            server.create('feature-toggle', {
+              certifPrescriptionSco: isFeatureToggleEnabled,
             });
+            const connectedUserId = isUserSco ? scoUser.id : user.id;
+            await authenticateSession(connectedUserId);
+
+            // when
+            await visit(`/sessions/${session.id}`);
+            await click(linkToCandidate);
+
+            // then
+            assert.equal(currentURL(), `/sessions/${session.id}/candidats`);
           });
         });
       });
     });
   });
+
+  async function _fillFormWithCorrectData() {
+    await _fillCandidateInput('lastName', validLastName);
+    await _fillCandidateInput('firstName', validFirstName);
+    await _fillCandidateInput('birthCity', validBirthCity);
+    await _fillCandidateInput('birthProvinceCode', validBirthProvinceCode);
+    await _fillCandidateInput('birthCountry', validBirthCountry);
+    await _fillCandidateInput('birthdate', validBirthdate);
+  }
+
+  async function _fillCandidateInput(code, value) {
+    await fillIn(`[data-test-id="panel-candidate__${code}__add-staging"] > div > input`, value);
+  }
+
 });
